requirements.txt
Flask>=2.0.0
icalendar>=4.0.0
requests>=2.25.0
PyYAML>=5.4.0
APScheduler>=3.7.0
python-dateutil>=2.8.0
python-decouple>=3.4.0